ApplyHerePro/requirements.txt
asgiref==3.5.2
astroid==3.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==38.0.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.2
idna==3.4
isort==5.12.0
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.2
Pillow==9.2.0
platformdirs==3.9.1
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.6.0
pylint==3.0.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
requirements.py==1.0.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomlkit==0.12.2
tzdata==2023.3
urllib3==1.26.12
virtualenv==20.24.0
